digfx.cpp : Add 16bpp case of RAW gfx layout drawgfx.cpp, tilemap.cpp : Make changeable total elements of gfx_elements constructor at RAW case, Allow 16bpp gfxdecode, Fix spacing tilemap.cpp : Allow 16bpp tilemap pen data gstream.cpp : Use gfxdecode for 16bpp gfx of X2222, Fix spacing igs017_igs031.cpp : Cleanup single sprite drawing routine jedi.cpp : Improve debug gfxdecode viewer(text gfx)

+++ src/devices/video/imagetek_i4100.cpp
@@ -1085,7 +1085,7 @@ void imagetek_i4100_device::draw_sprites( screen_device &screen, bitmap_ind16 &b
 		gfx_element *gfx1 = gfx(big ? 3 : 1);
 		uint32_t tile2 = big ? ((tile & 0xfffff) + ((code & 0xf) << 3)) :
 								((tile & 0xfffff) + ((code & 0xf) << 1));
-		const uint8_t* data;
+		const uint16_t* data;
 		uint8_t flipxy = (code & 0x6000) >> 13;
 
 		if (tile2 < gfx1->elements())
@@ -1117,7 +1117,7 @@ void imagetek_i4100_device::draw_sprites( screen_device &screen, bitmap_ind16 &b
 		gfx_element *gfx1 = gfx(big ? 2 : 0);
 		uint32_t tile2 = big ? ((tile & 0xfffff) + ((code & 0xf) << 2)) :
 								((tile & 0xfffff) + (code & 0xf));
-		const uint8_t* data;
+		const uint16_t* data;
 		uint8_t flipxy = (code & 0x6000) >> 13;
 
 		if (tile2 < gfx1->elements())
